requirements.txt
fastapi==0.116.1
sqlalchemy
uvicorn
pandas
numpy
psycopg2-binary
python-dotenv
